Added '**/*.vdi' to .gitignore to ignore VBox Disk Images
removed "Whiteout root" part between 28:34, since this is redundant with what lines 52:54 do

Added an if-clause around swap whiteout code to respect the case where no swap is created.
Added a for-loop around swap whiteout to respect multiple swaps if available.

+++ scripts/cleanup.sh
@@ -25,13 +25,6 @@ rm -rfv /usr/src/vboxguest* /usr/src/virtualbox-ose-guest*
 
 # Clean any Python bytecode
 find / -type f -name "*.pyc" -delete
-
-# Whiteout root
-echo "Whiteouting /"
-count=`df --sync -kP / | tail -n1  | awk -F ' ' '{print $4}'`;
-count=$((count -= 1))
-dd if=/dev/zero of=/tmp/whitespace bs=1024 count=$count;
-rm /tmp/whitespace;
  
 # Whiteout /boot
 echo "Whiteouting /boot"
@@ -41,12 +34,15 @@ dd if=/dev/zero of=/boot/whitespace bs=1024 count=$count;
 rm /boot/whitespace;
  
 # Whiteout swap 
-echo "Whiteouting swap"
-swappart=`cat /proc/swaps | tail -n1 | awk -F ' ' '{print $1}'`
-swapoff $swappart;
-dd if=/dev/zero of=$swappart;
-mkswap $swappart;
-swapon $swappart;
+echo "Whiteouting swaps"
+if [ $(cat /proc/swaps | wc -l) -gt 1 ]; then
+    for swappart in `cat /proc/swaps | tail -n +2 | awk -F ' ' '{print $1}'`; do
+        swapoff $swappart;
+        dd if=/dev/zero of=$swappart;
+        mkswap $swappart;
+        swapon $swappart;
+    done
+fi
 
 # Zero out the rest of the free space using dd.
 echo "Zero out the free space to aid VM compression"
